Move the new plugin to its own file
