+++ scripts/rollup/plugins/sizes-plugin.js
@@ -1,3 +1,11 @@
+/**
+ * Copyright (c) 2013-present, Facebook, Inc.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+'use strict';
+
 const gzip = require('gzip-size');
 
 module.exports = function sizes(options) {
